rename create_examples_py into create3_examples_py
