+++ create3_examples_py/create3_examples_py/dance/create3_dance.py
@@ -1,5 +1,5 @@
 import rclpy
-from create_examples_py.dance import dance_choreograph as dc
+from create3_examples_py.dance import dance_choreograph as dc
