+++ create3_examples_py/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/create3_examples_py
+[install]
+install_scripts=$base/lib/create3_examples_py
