+++ create3_examples_py/test/test_copyright.py
