--- create_examples_py/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/create_examples_py
-[install]
-install_scripts=$base/lib/create_examples_py
